d2xx: move out of experimental, add to hostextra.Init()
It's starting to take shape. While it is not yet fully stable, I'm moving it out
of experimental as it's starting to get into better shape and I'm relatively
happy with the API. Also by enabling hostextra.Init() to work right away, this
permits me to do other kinds of experimentations, like enabling gpio-list and
header-list to work really soon.