d2xx: export EEPROM
Export EEPROM. Reading work but writing doesn't, even with the Adafruit device
so it's a bug in the code, it's not a problem with the CJMCU sourced ones.

Prepare for late-initialization, which will make the periph.Init() call faster,
deferring work a bit. I'm not sure I'll use this in the end but the code is
better laid out.