d2xx: register in gpioreg
Sadly tools like gpio-read / gpio-write / gpio-list / headers-list aren't as
useful as they could, as the driver resets the state of the device when is
starts.

I'll try to start without resetting and reset when state is hard to figure out
later.

For FT232H, it'll probably be: poke if already in MPSSE with mpsseVerify(), if
so, assumes the device was correctly initialized and do not try initializing
again.