d2xx: FT232H: In() stop accepting Float
Have it accept the preconfigured pull instead, and return it in Pull().

In practice the FT232H goes through a journey of different I/O state when the
device starts up, and when it's in suspend, and its behavior can be changed in
the EEPROM.

Remove debugging code left over from 1be1a3ab5bea15683.